Fixes broken enum matching in token query msg builder; XXX TODO: implement this same fix for TokenExecuteMsg
