+++ src/main.rs
@@ -8,7 +8,6 @@ pub mod server;
 use rmcp::{ServiceExt, transport::stdio};
 use std::error::Error as StdError;
 
-// use crate::server::{ambur::AmburMcp, token::AmburTokenMcp};
 use crate::server::ambur::AmburMcp;
 
 #[tokio::main]
@@ -16,12 +15,8 @@ async fn main() -> Result<(), Box<dyn StdError>> {
     let ambur_mcp = AmburMcp::new().serve(stdio()).await.inspect_err(|e| {
         println!("{e}");
     })?;
-    ambur_mcp.waiting().await?;
 
-    // let token_mcp = AmburTokenMcp::new().serve(stdio()).await.inspect_err(|e| {
-    //     println!("{e}");
-    // })?;
-    // token_mcp.waiting().await?;
+    ambur_mcp.waiting().await?;
 
     Ok(())
 }
