+++ src/server/ambur.rs
@@ -11,6 +11,7 @@ use crate::contract::*;
 use crate::execute::*;
 use crate::instruction::*;
 use crate::network::*;
+use crate::server::token::*;
 use crate::query::{AllResponse as AllQueryResponse, ValidatedQuery};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
@@ -241,6 +242,170 @@ impl AmburMcp {
             contract_addr,
             msg: to_json_binary(&deserialized).unwrap_or_default(),
             funds,
+        }
+        .into();
+        let serialized_cosmos_msg = serde_json::to_string(&cosmos_msg);
+        if serialized_cosmos_msg.is_err() {
+            return Ok(CallToolResult::error(vec![Content::text(
+                "Error wrapping ExecuteMsg as CosmosMsg",
+            )]));
+        }
+        let valid_execute = ValidatedExecute {
+            execute_msg,
+            cosmos_msg: serialized_cosmos_msg.unwrap_or_default(),
+        };
+        let serialized: String = serde_json::to_string(&valid_execute).unwrap_or_default();
+        Ok(CallToolResult::success(vec![Content::text(serialized)]))
+    }
+
+    // cw721 Query entry point tools
+    #[tool(description = "XXX TODO")]
+    async fn list_token_query_entry_points(
+        &self,
+        #[tool(param)]
+        #[schemars(
+            description = "name of the NFT collection (e.g. \"archies\", \"the foresight ticket\", \"derpies\", \"ghouls\")"
+        )]
+        nft: String,
+    ) -> Result<CallToolResult, Error> {
+        let schema = match nft.to_lowercase().as_str() {
+            "archies" => Some(schema_for!(ArchiesQueryMsg<ArchiesExtension>)),
+            "derpies" => Some(schema_for!(DerpiesQueryMsg<DerpiesExtension>)),
+            "ghouls" => Some(schema_for!(GhoulsQueryMsg<GhoulsExtension>)),
+            "foresight" => Some(schema_for!(ForesightQueryMsg<ForesightExtension>)),
+            "the foresight ticket" => Some(schema_for!(ForesightQueryMsg<ForesightExtension>)),
+            _ => None,
+        };
+        if schema.is_none() {
+            let err_msg = "Unrecognized NFT collection name ".to_string() + &nft;
+            return Ok(CallToolResult::error(vec![Content::text(&err_msg)]));
+        }
+        let schema = schema.unwrap();
+        let serialized: String = serde_json::to_string(&schema).unwrap_or("".to_string());
+        Ok(CallToolResult::success(vec![Content::text(serialized)]))
+    }
+
+    #[tool(description = "XXX TODO")]
+    async fn build_token_query_msg(
+        &self,
+        #[tool(param)]
+        #[schemars(
+            description = "name of the NFT collection (e.g. \"archies\", \"the foresight ticket\", \"derpies\", \"ghouls\")"
+        )]
+        nft: String,
+        #[tool(param)]
+        #[schemars(description = "contract address of cw721 token")]
+        contract_addr: String,
+        #[tool(param)]
+        #[schemars(
+            description = "JSON stringified QueryMsg variant needed for building the query as a Cosmos SDK QueryRequest"
+        )]
+        query_msg: String,
+    ) -> Result<CallToolResult, Error> {
+        let query_req: Option<QueryRequest> = match nft.to_lowercase().as_str() {
+            "archies" => {
+                let deserialized: ArchiesQueryMsg<ArchiesExtension> = serde_json::from_str(query_msg.as_str()).unwrap();
+                let query_req = QueryRequest::Wasm(WasmQuery::Smart {
+                    contract_addr,
+                    msg: to_json_binary(&deserialized).unwrap_or_default(),
+                });
+                Some(query_req)
+            },
+            "derpies" => {
+                let deserialized: DerpiesQueryMsg<DerpiesExtension> = serde_json::from_str(query_msg.as_str()).unwrap();
+                let query_req = QueryRequest::Wasm(WasmQuery::Smart {
+                    contract_addr,
+                    msg: to_json_binary(&deserialized).unwrap_or_default(),
+                });
+                Some(query_req)
+            },
+            "ghouls" => {
+                let deserialized: GhoulsQueryMsg<GhoulsExtension> = serde_json::from_str(query_msg.as_str()).unwrap();
+                let query_req = QueryRequest::Wasm(WasmQuery::Smart {
+                    contract_addr,
+                    msg: to_json_binary(&deserialized).unwrap_or_default(),
+                });
+                Some(query_req)
+            },
+            "foresight" => {
+                let deserialized: ForesightQueryMsg<ForesightExtension> = serde_json::from_str(query_msg.as_str()).unwrap();
+                let query_req = QueryRequest::Wasm(WasmQuery::Smart {
+                    contract_addr,
+                    msg: to_json_binary(&deserialized).unwrap_or_default(),
+                });
+                Some(query_req)
+            },
+            "the foresight ticket" => {
+                let deserialized: ForesightQueryMsg<ForesightExtension> = serde_json::from_str(query_msg.as_str()).unwrap();
+                let query_req = QueryRequest::Wasm(WasmQuery::Smart {
+                    contract_addr,
+                    msg: to_json_binary(&deserialized).unwrap_or_default(),
+                });
+                Some(query_req)
+            },
+            _ => None,
+        };
+        if query_req.is_none() {
+            return Ok(CallToolResult::error(vec![Content::text(
+                "Error deserializing token QueryMsg",
+            )]));
+        }
+        let serialized_query_req = serde_json::to_string(&query_req);
+        if serialized_query_req.is_err() {
+            return Ok(CallToolResult::error(vec![Content::text(
+                "Error wrapping token QueryMsg as QueryRequest",
+            )]));
+        }
+        let valid_query = ValidatedQuery {
+            query_msg,
+            query_request: serialized_query_req.unwrap_or_default(),
+        };
+        let serialized: String = serde_json::to_string(&valid_query).unwrap_or_default();
+        Ok(CallToolResult::success(vec![Content::text(serialized)]))
+    }
+
+    // cw721 Execute entry point tools
+    #[tool(description = "XXX TODO")]
+    async fn list_token_tx_entry_points(
+        &self,
+        #[tool(param)]
+        #[schemars(
+            description = "name of the NFT collection (e.g. \"archies\", \"the foresight ticket\", \"derpies\", \"ghouls\")"
+        )]
+        nft: String,
+    ) -> Result<CallToolResult, Error> {
+        let schema = match nft.to_lowercase().as_str() {
+            "archies" => Some(schema_for!(ArchiesExecuteMsg)),
+            "derpies" => Some(schema_for!(DerpiesExecuteMsg)),
+            "ghouls" => Some(schema_for!(GhoulsExecuteMsg)),
+            "foresight" => Some(schema_for!(ForesightExecuteMsg)),
+            "the foresight ticket" => Some(schema_for!(ForesightExecuteMsg)),
+            _ => None,
+        };
+        if schema.is_none() {
+            let err_msg = "Unrecognized NFT collection name ".to_string() + &nft;
+            return Ok(CallToolResult::error(vec![Content::text(&err_msg)]));
+        }
+        let schema = schema.unwrap();
+        let serialized: String = serde_json::to_string(&schema).unwrap_or("".to_string());
+        Ok(CallToolResult::success(vec![Content::text(serialized)]))
+    }
+
+    #[tool(description = "XXX TODO")]
+    async fn build_token_execute_msg(
+        &self,
+        #[tool(param)]
+        #[schemars(description = "XXX TODO")]
+        contract_addr: String,
+        #[tool(param)]
+        #[schemars(description = "XXX TODO")]
+        execute_msg: String,
+    ) -> Result<CallToolResult, Error> {
+        let deserialized: TokenExecuteMsg = serde_json::from_str(execute_msg.as_str()).unwrap();
+        let cosmos_msg: CosmosMsg = WasmMsg::Execute {
+            contract_addr,
+            msg: to_json_binary(&deserialized).unwrap_or_default(),
+            funds: vec![],
         }
         .into();
         let serialized_cosmos_msg = serde_json::to_string(&cosmos_msg);
